requirements.txt
pybit
websocket-client
kivy
pandas
numpy
requests
optuna
